Netrw. Make netrw's mc/mm/D working
Default option parameters such as g:netrw_localcoppycmd
and others overwrited the values, that were set
in _vimrc. So use default values for the parameters
if the parameters are not set in _vimrc

Removed extra `'` quotes in the copy system call.
Couldn't find a reason for these quotes. There is
a `s:ShellEscape` method that adds quotes and etc.

Replaced `delete(rmfile,"d")` with `delete(rmfile,"rf")`
for the netrw-D command, so as the empty folder can
be removed as well.